bugfixed for windows according to mathworks:explicit conversion for sqrt/floor, magma_free instead of free, no headers included
git-svn-id: https://icl.cs.utk.edu/svn/magma/trunk@6151 02bb5f2a-90ca-11de-bb24-8aedcc835bc8

+++ sparse-iter/blas/zgeellrtmv.cu
@@ -275,7 +275,7 @@ magma_zgeellrtmv(
     if ( arch < 200 && num_threads > 256 )
         printf("error: too much shared memory requested.\n");
 
-    int dimgrid1 = sqrt(num_blocks);
+    int dimgrid1 = (int) sqrt( (double) num_blocks );
     int dimgrid2 = (num_blocks + dimgrid1 -1 ) / dimgrid1;
     dim3 grid( dimgrid1, dimgrid2, 1);
