+++ sparse-iter/blas/zlobpcg_shift.cu
@@ -97,7 +97,7 @@ magma_zlobpcg_shift(
 
     dim3 block( num_threads, 1, 1 );
 
-    int dimgrid1 = sqrt(num_rows);
+    int dimgrid1 = (int) sqrt( (double) num_rows);
     int dimgrid2 = (num_rows + dimgrid1 -1 ) / dimgrid1;
 
     dim3 grid( dimgrid1, dimgrid2, 1);
